My tangent has completed!
Seriously though, I cleaned up the code a bit,
figured out what was causing the segmentation fault
(which happened to be me being stupid with linked lists again),
and we have a fully working example of the hotplug API ideas.

I'm pleased to say it handles failed connection callbacks,
but unfortunately, I had to make new allocations for each
node that's held in the linked list of devices for each driver.
They aren't large allocations but still a bummer.